[Silver V] Title: 크리 문자열, Time: 156 ms, Memory: 31120 KB -BaekjoonHub

+++ 백준/Silver/11059. 크리 문자열/크리 문자열.py
@@ -1,7 +1,7 @@
 def check(li, n):
     return sum(li[:n]) == sum(li[n:])
 
-S=[*map(int, list(input()))]
+S=[*map(float, list(input()))]
 n=len(S)
 i=n-n%2
 flag = False
